Feat: 모듈 변경, 사진 초기화 기능 구현
