+++ PlanJ/app/src/main/java/com/boostcamp/planj/data/di/LogInModule.kt
@@ -1,10 +1,19 @@
 package com.boostcamp.planj.data.di
 
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStoreFile
+import androidx.room.Room
 import com.boostcamp.planj.BuildConfig
+import com.boostcamp.planj.data.db.AlarmInfoDao
+import com.boostcamp.planj.data.db.AppDatabase
 import com.boostcamp.planj.data.network.LoginApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
@@ -51,4 +60,27 @@ object LogInModule {
         return retrofit.create(LoginApi::class.java)
     }
 
+
+    //room
+    @Singleton
+    @Provides
+    fun provideRoom(@ApplicationContext context: Context): AppDatabase {
+        return Room.databaseBuilder(
+            context,
+            AppDatabase::class.java,
+            "planj"
+        ).build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideAlarmInfoDao(appDatabase: AppDatabase): AlarmInfoDao = appDatabase.alarmInfoDao()
+
+    //DataStore
+    @Singleton
+    @Provides
+    fun providePreferencesDataStore(@ApplicationContext context: Context): DataStore<Preferences> =
+        PreferenceDataStoreFactory.create(
+            produceFile = { context.preferencesDataStoreFile(BuildConfig.DATA_STORE_NAME) }
+        )
 }
