+++ PlanJ/app/src/main/java/com/boostcamp/planj/data/di/MainModule.kt
@@ -1,43 +1,28 @@
 package com.boostcamp.planj.data.di
 
-import android.content.Context
 import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.stringPreferencesKey
-import androidx.datastore.preferences.preferencesDataStoreFile
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import androidx.sqlite.db.SupportSQLiteDatabase
 import com.boostcamp.planj.BuildConfig
-import com.boostcamp.planj.data.db.AlarmInfoDao
-import com.boostcamp.planj.data.db.AppDatabase
-import com.boostcamp.planj.data.model.Category
 import com.boostcamp.planj.data.network.MainApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
+import dagger.hilt.android.components.ActivityRetainedComponent
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
 
 @Module
-@InstallIn(SingletonComponent::class)
+@InstallIn(ActivityRetainedComponent::class)
 object MainModule {
 
 
-    @Singleton
     @Provides
     fun provideInterceptor(datastore: DataStore<Preferences>): Interceptor {
         val user = stringPreferencesKey(BuildConfig.USER)
@@ -60,7 +45,6 @@ object MainModule {
     }
 
     //retrofit
-    @Singleton
     @Provides
     fun provideOkHttpClient(interceptor: Interceptor): OkHttpClient {
         val httpLoggingInterceptor = HttpLoggingInterceptor()
@@ -71,7 +55,6 @@ object MainModule {
             .build()
     }
 
-    @Singleton
     @Provides
     fun provideRetrofitInstance(okHttpClient: OkHttpClient): Retrofit {
         return Retrofit.Builder()
@@ -81,35 +64,10 @@ object MainModule {
             .build()
     }
 
-    @Singleton
     @Provides
     fun provideService(retrofit: Retrofit): MainApi {
         return retrofit.create(MainApi::class.java)
     }
 
-    //okhttp
-
-    //room
-    @Singleton
-    @Provides
-    fun provideRoom(@ApplicationContext context: Context): AppDatabase {
-        return Room.databaseBuilder(
-            context,
-            AppDatabase::class.java,
-            "planj"
-        ).build()
-    }
-
-    @Singleton
-    @Provides
-    fun provideAlarmInfoDao(appDatabase: AppDatabase): AlarmInfoDao = appDatabase.alarmInfoDao()
-
-    //DataStore
-    @Singleton
-    @Provides
-    fun providePreferencesDataStore(@ApplicationContext context: Context): DataStore<Preferences> =
-        PreferenceDataStoreFactory.create(
-            produceFile = { context.preferencesDataStoreFile(BuildConfig.DATA_STORE_NAME) }
-        )
 
 }
