+++ PlanJ/app/src/main/java/com/boostcamp/planj/data/di/NaverDirectionModule.kt
@@ -10,6 +10,8 @@ import com.boostcamp.planj.data.network.NaverApi
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ActivityComponent
+import dagger.hilt.android.components.ActivityRetainedComponent
 import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
