+++ PlanJ/app/src/main/java/com/boostcamp/planj/data/repository/LoginRepositoryImpl.kt
@@ -6,6 +6,9 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.core.stringPreferencesKey
+import com.boostcamp.planj.data.db.AlarmInfoDao
+import com.boostcamp.planj.data.model.AlarmInfo
+import com.boostcamp.planj.data.model.DateTime
 import com.boostcamp.planj.data.model.dto.SignInNaverRequest
 import com.boostcamp.planj.data.model.dto.LoginResponse
 import com.boostcamp.planj.data.model.dto.SignInRequest
@@ -18,11 +21,13 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
 import java.io.IOException
+import java.util.Calendar
 import javax.inject.Inject
 
 class LoginRepositoryImpl @Inject constructor(
     private val apiService: LoginApi,
-    private val dataStore: DataStore<Preferences>
+    private val dataStore: DataStore<Preferences>,
+    private val alarmInfoDao: AlarmInfoDao,
 ) : LoginRepository {
 
     companion object {
@@ -57,11 +62,11 @@ class LoginRepositoryImpl @Inject constructor(
         dataStore.edit { prefs ->
             prefs[USER] = id
         }
-        Log.d("PLANJDEBUG", "dataStore 저장 완료 ${getUser().first()}")
+        Log.d("PLANJDEBUG", "dataStore 저장 완료 ${getToken().first()}")
     }
 
 
-    override fun getUser(): Flow<String> {
+    override fun getToken(): Flow<String> {
         return dataStore.data
             .catch { e ->
                 if (e is IOException) {
@@ -78,5 +83,89 @@ class LoginRepositoryImpl @Inject constructor(
 
     override fun postSignInNaver(accessToken: String): Flow<LoginResponse> = flow {
         emit(apiService.postSignInNaver(SignInNaverRequest(accessToken)))
+    }
+
+    override suspend fun updateAlarmInfo(curTimeMillis: Long) {
+        val alarmList = alarmInfoDao.getAll()
+        val calendar = Calendar.getInstance()
+        alarmList.forEach { alarmInfo ->
+            calendar.timeInMillis = alarmInfo.endTime.toMilliseconds()
+            calendar.add(Calendar.MINUTE, -alarmInfo.alarm.alarmTime - alarmInfo.estimatedTime)
+
+            // 알림 시간 < 현재 시간 -> 알람 삭제 or 업데이트
+            if (calendar.timeInMillis < curTimeMillis) {
+                // 일회성 -> 알림 DB에서 삭제
+                // 반복 -> DB 업데이트
+                if (alarmInfo.repetition == null) {
+                    alarmInfoDao.deleteAlarmInfo(alarmInfo)
+                } else {
+                    val oneDayMillis = 24 * 60 * 60 * 1000L
+                    val interval = if (alarmInfo.repetition.cycleType == "DAILY") {
+                        alarmInfo.repetition.cycleCount
+                    } else {
+                        alarmInfo.repetition.cycleCount * 7
+                    }
+                    while (calendar.timeInMillis < curTimeMillis) {
+                        calendar.timeInMillis += (interval * oneDayMillis)
+                    }
+                    calendar.add(
+                        Calendar.MINUTE,
+                        alarmInfo.alarm.alarmTime + alarmInfo.estimatedTime
+                    )
+                    val endTime = DateTime(
+                        calendar.get(Calendar.YEAR),
+                        calendar.get(Calendar.MONTH) + 1,
+                        calendar.get(Calendar.DAY_OF_MONTH),
+                        calendar.get(Calendar.HOUR_OF_DAY),
+                        calendar.get(Calendar.MINUTE)
+                    )
+                    alarmInfoDao.updateAlarmInfo(
+                        alarmInfo.copy(endTime = endTime)
+                    )
+                }
+            }
+        }
+    }
+
+    override suspend fun getAllAlarmInfo(): List<AlarmInfo> {
+        return alarmInfoDao.getAll()
+    }
+
+    override suspend fun deleteAllData() {
+        alarmInfoDao.deleteAllAlarmInfo()
+    }
+
+    override suspend fun insertAlarmInfo(alarmInfo: AlarmInfo) {
+        alarmInfoDao.insertAlarmInfo(alarmInfo)
+    }
+
+    override suspend fun deleteAlarmInfo(alarmInfo: AlarmInfo) {
+        alarmInfoDao.deleteAlarmInfo(alarmInfo)
+    }
+
+    override suspend fun deleteAlarmInfoUsingScheduleId(scheduleId: String) {
+        alarmInfoDao.deleteAlarmInfoUsingScheduleId(scheduleId)
+    }
+
+    override suspend fun saveAlarmMode(mode: Boolean) {
+        dataStore.edit { prefs ->
+            prefs[MainRepositoryImpl.ALARM_MODE] = mode
+        }
+    }
+
+    override fun getAlarmMode(): Flow<Boolean> {
+        return dataStore.data
+            .catch { exception ->
+                if (exception is IOException) {
+                    exception.printStackTrace()
+                    emit(emptyPreferences())
+                } else {
+                    throw exception
+                }
+
+            }
+            .map { prefs ->
+                prefs[MainRepositoryImpl.ALARM_MODE] ?: true
+            }
     }
 }
