+++ PlanJ/app/src/main/java/com/boostcamp/planj/data/repository/MainRepositoryImpl.kt
@@ -41,7 +41,6 @@ import javax.inject.Inject
 class MainRepositoryImpl @Inject constructor(
     private val api: MainApi,
     private val dataStore: DataStore<Preferences>,
-    private val alarmInfoDao: AlarmInfoDao,
 ) : MainRepository {
 
     companion object {
@@ -65,23 +64,8 @@ class MainRepositoryImpl @Inject constructor(
 
     override suspend fun emptyToken() {
         dataStore.edit { prefs ->
-            prefs[USER] = ""
+            prefs.clear()
         }
-    }
-
-    override fun getToken(): Flow<String> {
-        return dataStore.data
-            .catch { e ->
-                if (e is IOException) {
-                    e.printStackTrace()
-                    emit(emptyPreferences())
-                } else {
-                    throw e
-                }
-            }
-            .map { pref ->
-                pref[USER] ?: ""
-            }
     }
 
     override suspend fun deleteScheduleApi(scheduleUuid: String) {
@@ -112,40 +96,42 @@ class MainRepositoryImpl @Inject constructor(
 
     override suspend fun getCategorySchedulesApi(categoryUuid: String): Flow<GetSchedulesResponse> =
         flow {
-            try {
-                val scheduleInfo = api.getCategorySchedule(categoryUuid)
-                val scheduleDummy = scheduleInfo.date.map {
-
-                    val startAt =
-                        it.startAt?.split("T", "-", ":")?.map { time -> time.toInt() } ?: emptyList()
-                    val endAt = it.endAt.split("T", "-", ":").map { time -> time.toInt() }
-
-                    Schedule(
-                        scheduleId = it.scheduleUuid,
-                        title = it.title,
-                        startAt = if (startAt.isEmpty()) null else DateTime(
-                            startAt[0],
-                            startAt[1],
-                            startAt[2],
-                            startAt[3],
-                            startAt[4],
-                            startAt[5]
-                        ),
-                        endAt = DateTime(endAt[0], endAt[1], endAt[2], endAt[3], endAt[4], endAt[5]),
-                        isFinished = it.isFinished,
-                        isFailed = it.isFailed,
-                        repeated = it.repeated,
-                        hasRetrospectiveMemo = it.hasRetrospectiveMemo,
-                        shared = it.shared,
-                        participantCount = it.participantCount,
-                        participantSuccessCount = it.participantSuccessCount
-                    )
-                }
-                emit(GetSchedulesResponse("",scheduleDummy))
-
-            } catch (e: Exception) {
-                Log.d("PLANJDEBUG", "getCategorySchedulesApi error  ${e.message}")
-            }
+//            try {
+//                val scheduleInfo = api.getCategorySchedule(categoryUuid)
+//                val scheduleDummy = scheduleInfo.date.map {
+//
+//                    Log.d("PLANJDEBUG", "${it.startAt}, ${it.endAt}")
+//
+//                    val startAt =
+//                        it.startAt?.split("T", "-", ":")?.map { time -> time.toInt() } ?: emptyList()
+//                    val endAt = it.endAt.split("T", "-", ":").map { time -> time.toInt() }
+//                    Schedule(
+//                        scheduleId = it.scheduleUuid,
+//                        title = it.title,
+//                        startAt = if (startAt.isEmpty()) null else DateTime(
+//                            startAt[0],
+//                            startAt[1],
+//                            startAt[2],
+//                            startAt[3],
+//                            startAt[4],
+//                            startAt[5]
+//                        ),
+//                        endAt = DateTime(endAt[0], endAt[1], endAt[2], endAt[3], endAt[4], endAt[5]),
+//                        isFinished = it.isFinished,
+//                        isFailed = it.isFailed,
+//                        repeated = it.repeated,
+//                        hasRetrospectiveMemo = it.hasRetrospectiveMemo,
+//                        shared = it.shared,
+//                        participantCount = it.participantCount,
+//                        participantSuccessCount = it.participantSuccessCount
+//                    )
+//                }
+//                emit(scheduleDummy)
+//
+//            } catch (e: Exception) {
+//                Log.d("PLANJDEBUG", "getCategorySchedulesApi error ${e.message}")
+//            }
+            emit(api.getCategorySchedule(categoryUuid))
         }
 
     override suspend fun getWeeklyScheduleApi(date: String): Flow<GetSchedulesResponse> = flow {
@@ -212,89 +198,6 @@ class MainRepositoryImpl @Inject constructor(
         emit(api.patchUser(nickName, imageFile))
     }
 
-    override suspend fun saveAlarmMode(mode: Boolean) {
-        dataStore.edit { prefs ->
-            prefs[ALARM_MODE] = mode
-        }
-    }
-
-    override suspend fun getAlarmMode(): Flow<Boolean> {
-        return dataStore.data
-            .catch { exception ->
-                if (exception is IOException) {
-                    exception.printStackTrace()
-                    emit(emptyPreferences())
-                } else {
-                    throw exception
-                }
-
-            }
-            .map { prefs ->
-                prefs[ALARM_MODE] ?: true
-            }
-    }
-
-    override suspend fun deleteAllData() {
-        alarmInfoDao.deleteAllAlarmInfo()
-    }
-
-    override suspend fun insertAlarmInfo(alarmInfo: AlarmInfo) {
-        alarmInfoDao.insertAlarmInfo(alarmInfo)
-    }
-
-    override suspend fun getAllAlarmInfo(): List<AlarmInfo> {
-        return alarmInfoDao.getAll()
-    }
-
-    override suspend fun deleteAlarmInfo(alarmInfo: AlarmInfo) {
-        alarmInfoDao.deleteAlarmInfo(alarmInfo)
-    }
-
-    override suspend fun deleteAlarmInfoUsingScheduleId(scheduleId: String) {
-        alarmInfoDao.deleteAlarmInfoUsingScheduleId(scheduleId)
-    }
-
-    override suspend fun updateAlarmInfo(curTimeMillis: Long) {
-        val alarmList = alarmInfoDao.getAll()
-        val calendar = Calendar.getInstance()
-        alarmList.forEach { alarmInfo ->
-            calendar.timeInMillis = alarmInfo.endTime.toMilliseconds()
-            calendar.add(Calendar.MINUTE, -alarmInfo.alarm.alarmTime - alarmInfo.estimatedTime)
-
-            // 알림 시간 < 현재 시간 -> 알람 삭제 or 업데이트
-            if (calendar.timeInMillis < curTimeMillis) {
-                // 일회성 -> 알림 DB에서 삭제
-                // 반복 -> DB 업데이트
-                if (alarmInfo.repetition == null) {
-                    alarmInfoDao.deleteAlarmInfo(alarmInfo)
-                } else {
-                    val oneDayMillis = 24 * 60 * 60 * 1000L
-                    val interval = if (alarmInfo.repetition.cycleType == "DAILY") {
-                        alarmInfo.repetition.cycleCount
-                    } else {
-                        alarmInfo.repetition.cycleCount * 7
-                    }
-                    while (calendar.timeInMillis < curTimeMillis) {
-                        calendar.timeInMillis += (interval * oneDayMillis)
-                    }
-                    calendar.add(
-                        Calendar.MINUTE,
-                        alarmInfo.alarm.alarmTime + alarmInfo.estimatedTime
-                    )
-                    val endTime = DateTime(
-                        calendar.get(Calendar.YEAR),
-                        calendar.get(Calendar.MONTH) + 1,
-                        calendar.get(Calendar.DAY_OF_MONTH),
-                        calendar.get(Calendar.HOUR_OF_DAY),
-                        calendar.get(Calendar.MINUTE)
-                    )
-                    alarmInfoDao.updateAlarmInfo(
-                        alarmInfo.copy(endTime = endTime)
-                    )
-                }
-            }
-        }
-    }
 
     override suspend fun getDetailSchedule(scheduleId: String): ScheduleDetail {
         return api.getDetailSchedule(scheduleId).scheduleDetail
@@ -303,5 +206,8 @@ class MainRepositoryImpl @Inject constructor(
     override fun getScheduleChecked(scheduleId: String): Flow<GetScheduleCheckedResponse> =
         flow { emit(api.getScheduleChecked(scheduleId)) }
 
+    override suspend fun getUserImageRemove() {
+        return api.patchUserImageRemove()
+    }
 }
