+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/RestartAlarmReceiver.kt
@@ -3,6 +3,7 @@ package com.boostcamp.planj.ui
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
+import com.boostcamp.planj.data.repository.LoginRepository
 import com.boostcamp.planj.data.repository.MainRepository
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
@@ -14,14 +15,14 @@ import javax.inject.Inject
 class RestartAlarmReceiver : BroadcastReceiver() {
 
     @Inject
-    lateinit var mainRepository: MainRepository
+    lateinit var loginRepository: LoginRepository
 
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action == "android.intent.action.BOOT_COMPLETED") {
             val alarm = PlanjAlarm(context)
             CoroutineScope(Dispatchers.IO).launch {
-                mainRepository.updateAlarmInfo(System.currentTimeMillis())
-                val alarmList = mainRepository.getAllAlarmInfo()
+                loginRepository.updateAlarmInfo(System.currentTimeMillis())
+                val alarmList = loginRepository.getAllAlarmInfo()
                 alarmList.forEach { alarmInfo -> alarm.setAlarm(alarmInfo) }
             }
         }
