+++ PlanJ/app/src/main/res/layout/fragment_setting.xml
@@ -99,6 +99,17 @@
                     app:layout_constraintTop_toTopOf="parent" />
 
                 <ImageView
+                    android:id="@+id/iv_setting_icon_cancel"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:contentDescription="@string/edit_profile"
+                    android:src="@drawable/ic_cancel"
+                    android:visibility="@{viewModel.isEditMode() ? View.VISIBLE : View.GONE}"
+                    app:layout_constraintTop_toTopOf="@id/iv_setting_img"
+                    app:layout_constraintEnd_toEndOf="@id/iv_setting_img" />
+
+
+                <ImageView
                     android:id="@+id/iv_setting_icon_camera"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -107,6 +118,8 @@
                     android:visibility="@{viewModel.isEditMode() ? View.VISIBLE : View.GONE}"
                     app:layout_constraintBottom_toBottomOf="@id/iv_setting_img"
                     app:layout_constraintEnd_toEndOf="@id/iv_setting_img" />
+
+
 
                 <EditText
                     android:id="@+id/tv_setting_nickname"
