+++ PlanJ/app/src/main/res/navigation/main_nav.xml
@@ -23,7 +23,10 @@
         >
         <action
             android:id="@+id/action_fragment_user_to_loginActivity"
-            app:destination="@id/loginActivity" />
+            app:destination="@id/loginActivity"
+            app:popUpTo="@layout/fragment_home"
+            app:popUpToInclusive="true"
+            />
     </fragment>
     <fragment
         android:id="@+id/fragment_friend_list"
